+++ legacy/src/main/java/org/openscience/cdk/aromaticity/DoubleBondAcceptingAromaticityDetector.java
@@ -177,7 +177,7 @@ public class DoubleBondAcceptingAromaticityDetector {
     }
 
     private static int getLonePairCount(IAtom atom) {
-        Integer count = (Integer) atom.getProperty(CDKConstants.LONE_PAIR_COUNT);
+        Integer count = atom.getProperty(CDKConstants.LONE_PAIR_COUNT);
         if (count == null) {
             return 0;
         } else {
